Abstracted some methods from ray_benchmark into functions for use in tests
