+++ tools/CMakeLists.txt
@@ -34,26 +34,4 @@ foreach(tool ${TOOL_NAMES})
     install(TARGETS ${tool_exec} DESTINATION ${CMAKE_INSTALL_BINDIR}/tools)
 endforeach()
 
-#===============================================================================
-# ray-benchmark (special case - requires linking directly to GPRT)
-#===============================================================================
-if (XDG_ENABLE_GPRT)
-    # Embed and compile the device code
-    embed_devicecode(
-        OUTPUT_TARGET
-        ray_benchmark_deviceCode
-        HEADERS
-        ${CMAKE_CURRENT_SOURCE_DIR}/ray_benchmark_shared.h
-        SOURCES
-        ${CMAKE_CURRENT_SOURCE_DIR}/ray_benchmark_deviceCode.slang
-    )
-
-    # Create the ray-benchmark executable
-    add_executable(ray-benchmark ray_benchmark.cpp)
-    target_link_libraries(ray-benchmark xdg argparse ray_benchmark_deviceCode)
-    if (OpenMP_CXX_FOUND)
-        target_link_libraries(ray-benchmark OpenMP::OpenMP_CXX)
-        target_compile_definitions(ray-benchmark PUBLIC XDG_OPENMP)
-    endif()
-    install(TARGETS ray-benchmark DESTINATION ${CMAKE_INSTALL_BINDIR}/tools)
-endif()
+add_subdirectory(ray_benchmark)
