+++ tools/ray_benchmark/ray_benchmark.cpp
@@ -3,7 +3,6 @@
 #include <string>
 #include <vector>
 #include <random>
-#include <cmath>
 
 #include "xdg/error.h"
 #include "xdg/mesh_manager_interface.h"
@@ -15,47 +14,11 @@
 
 #include "argparse/argparse.hpp"
 
-// GPRT includes - only for use in ray generation callback
-#include "gprt/gprt.h"
-#include "ray_benchmark_shared.h"
+#include "ray_benchmark.h"
 
 #include <omp.h>
 
 using namespace xdg;
-extern GPRTProgram ray_benchmark_deviceCode;
-
-inline double rand01(uint32_t &state)
-{
-  state = state * 1664525u + 1013904223u;
-  return double(state) * (1.0 / 4294967296.0);
-}
-
-inline Direction random_unit_dir_lcg(uint32_t &state)
-{
-  double x1, x2, s;
-  do {
-    x1 = rand01(state) * 2.0 - 1.0;
-    x2 = rand01(state) * 2.0 - 1.0;
-    s  = x1 * x1 + x2 * x2;
-  } while (s <= 0.0 || s >= 1.0);
-
-  double t = 2.0 * std::sqrt(1.0 - s);
-  return { x1 * t, x2 * t, 1.0 - 2.0 * s };
-}
-
-// Generates a random point cloud with radius (--source-radius) 
-inline std::pair<Position,Direction> random_spherical_source(const Position& origin, std::uint32_t state, double source_radius)
-{
-  // Always generate random direction
-  Direction dir = random_unit_dir_lcg(state);
-  Position pos = origin;
-  if (source_radius > 0.0) {
-    // random origins (spherical source)
-    double r = source_radius * std::cbrt(rand01(state)); // uniform in ball
-    pos += dir * r;
-  } 
-  return {pos, dir};
-}
 
 int main(int argc, char** argv) {
 
@@ -187,44 +150,9 @@ int main(int argc, char** argv) {
     // ---- Random ray generation on device via callback method ----
     gen_timer.start();
 
-     
-    //  - User creates their own GPU compute API method to populate rays and passes that to XDG
-    //  - In this miniapp we are using GPRT as a demonstration
-    //  - This callback runs inside populate_rays_external and receives XDG's device buffers
     gprt_rt = std::dynamic_pointer_cast<GPRTRayTracer>(xdg->ray_tracing_interface());
-    if (!gprt_rt) {
-      fatal_error("GPRT backend requested but ray tracing interface is not GPRTRayTracer");
-    }
-
-    auto generateRaysCallback = [&](const DeviceRayHitBuffers& buffer, size_t numRays) {
-      GPRTContext context = gprt_rt->context();
-      GPRTModule module = gprtModuleCreate(context, ray_benchmark_deviceCode);
-      auto genRandomRays = gprtComputeCreate<GenerateRandomRayParams>(
-                           context, module, "generate_random_rays");
-
-      constexpr int threadsPerGroup = 64;
-      const int neededGroups = (int)((numRays + threadsPerGroup - 1) / threadsPerGroup);
-      const int groups = std::min(neededGroups, WORKGROUP_LIMIT);
-
-      GenerateRandomRayParams randomRayParams = {};
-      randomRayParams.rays = static_cast<dblRay*>(buffer.rayDevPtr); // Cast opaque pointer to typed dblRay*
-      randomRayParams.numRays = (uint32_t)numRays;
-      randomRayParams.source_radius = source_radius;
-      randomRayParams.origin = { origin.x, origin.y, origin.z };
-      randomRayParams.seed = seed;
-      randomRayParams.total_threads = (uint32_t)(groups * threadsPerGroup);
-      randomRayParams.volume_mesh_id = volume;
-      randomRayParams.enabled = 1u;
-
-      gprtComputeLaunch(genRandomRays,
-                        { (uint32_t)groups, 1, 1 },
-                        { (uint32_t)threadsPerGroup, 1, 1 },
-                        randomRayParams);
-      gprtComputeSynchronize(context);
-
-      gprtComputeDestroy(genRandomRays);
-      gprtModuleDestroy(module);
-    };
+    auto generateRaysCallback = 
+      tools::benchmark::make_generate_rays_callback(gprt_rt->context(), origin, source_radius, seed, volume);
 
     // Let XDG internally allocate buffers and invoke the callback to populate them
     xdg->populate_rays_external(N, generateRaysCallback);
@@ -250,7 +178,7 @@ int main(int argc, char** argv) {
     #pragma omp parallel for schedule(static)
     for (uint32_t i = 0; i < N; ++i) {
       uint32_t state = seed ^ i;
-      auto [pos,dir] = random_spherical_source(origin, state, source_radius);
+      auto [pos,dir] = tools::benchmark::random_spherical_source(origin, state, source_radius);
       origins[i] = pos;
       directions[i] = dir;
     }
