+++ tools/ray_benchmark/ray_benchmark_shared.h
@@ -1,0 +1,14 @@
+#include "gprt.h"
+
+#include "../../include/xdg/gprt/ray.h"
+
+struct GenerateRandomRayParams {
+  xdg::dblRay* rays;
+  uint numRays;
+  double3 origin;
+  uint seed;
+  uint total_threads;
+  double source_radius;
+  int volume_mesh_id;
+  uint enabled;
+};
